requirements.txt
# Core packages
llama_index>=0.9.0
openai>=1.2.0
python-dotenv>=1.0.0

# Document processing
pypdf>=3.16.0
docx2txt>=0.8
beautifulsoup4>=4.12.2
markdown>=3.5

# Utilities
tqdm>=4.66.1
requests>=2.31.0